requirements.txt
streamlit
langchain
langchain-openai
langchain-community
pandas
plotly
fpdf2
passlib
langchain-groq
python-dotenv
